Cleanup sdl input and fix xinput
