+++ Windows/Doom64EXPlus.vcxproj.filters
@@ -288,6 +288,9 @@
     <ClCompile Include="..\src\engine\deh_weapon.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\engine\i_sdlinput.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\src\engine\dgl.h">
@@ -510,6 +513,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\src\engine\z_zone.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\engine\i_sdlinput.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
