+++ src/engine/build_win_cross.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 i686-w64-mingw32.static-windres -O coff -i doom64ex-plus.rc -o doom64ex-plus.res
-i686-w64-mingw32.static-gcc `i686-w64-mingw32.static-pkg-config --cflags sdl2` -O2 -DWIN32 -D_WIN32 -D_USE_XINPUT -DWINDOWS -DWIN32_LEAN_AND_MEAN am_draw.c am_map.c info.c md5.c tables.c con_console.c con_cvar.c d_devstat.c d_main.c d_net.c f_finale.c in_stuff.c g_actions.c g_demo.c g_game.c g_settings.c wi_stuff.c m_cheat.c m_menu.c m_misc.c m_fixed.c m_keys.c m_password.c m_random.c m_shift.c net_client.c net_common.c net_dedicated.c net_io.c net_loop.c net_packet.c net_query.c net_server.c net_structrw.c dgl.c gl_draw.c gl_main.c gl_texture.c sc_main.c p_ceilng.c p_doors.c p_enemy.c p_user.c p_floor.c p_inter.c p_lights.c p_macros.c p_map.c p_maputl.c p_mobj.c p_plats.c p_pspr.c p_saveg.c p_setup.c p_sight.c p_spec.c p_switch.c p_telept.c p_tick.c r_clipper.c r_drawlist.c r_lights.c r_main.c r_scene.c r_bsp.c r_sky.c r_things.c r_wipe.c s_sound.c st_stuff.c i_audio.c i_main.c i_png.c i_system.c i_video.c w_file.c w_merge.c w_wad.c z_zone.c i_xinput.c -o DOOM64EX-Plus.exe `i686-w64-mingw32.static-pkg-config --libs sdl2` `i686-w64-mingw32.static-pkg-config --libs libpng` `i686-w64-mingw32.static-pkg-config --libs fluidsynth` `i686-w64-mingw32.static-pkg-config --libs gl` `i686-w64-mingw32.static-pkg-config --libs glu` `i686-w64-mingw32.static-pkg-config --libs SDL2_net` -lm doom64ex-plus.res
+i686-w64-mingw32.static-gcc `i686-w64-mingw32.static-pkg-config --cflags sdl2` -O2 -DWIN32 -D_WIN32 -D_USE_XINPUT -DWINDOWS -DWIN32_LEAN_AND_MEAN am_draw.c am_map.c info.c md5.c tables.c con_console.c con_cvar.c d_devstat.c d_main.c d_net.c f_finale.c in_stuff.c g_actions.c g_demo.c g_game.c g_settings.c wi_stuff.c m_cheat.c m_menu.c m_misc.c m_fixed.c m_keys.c m_password.c m_random.c m_shift.c net_client.c net_common.c net_dedicated.c net_io.c net_loop.c net_packet.c net_query.c net_server.c net_structrw.c dgl.c gl_draw.c gl_main.c gl_texture.c sc_main.c p_ceilng.c p_doors.c p_enemy.c p_user.c p_floor.c p_inter.c p_lights.c p_macros.c p_map.c p_maputl.c p_mobj.c p_plats.c p_pspr.c p_saveg.c p_setup.c p_sight.c p_spec.c p_switch.c p_telept.c p_tick.c r_clipper.c r_drawlist.c r_lights.c r_main.c r_scene.c r_bsp.c r_sky.c r_things.c r_wipe.c s_sound.c st_stuff.c i_audio.c i_main.c i_png.c i_system.c i_video.c w_file.c w_merge.c w_wad.c z_zone.c i_xinput.c i_sdlinput.c -o DOOM64EX-Plus.exe `i686-w64-mingw32.static-pkg-config --libs sdl2` `i686-w64-mingw32.static-pkg-config --libs libpng` `i686-w64-mingw32.static-pkg-config --libs fluidsynth` `i686-w64-mingw32.static-pkg-config --libs gl` `i686-w64-mingw32.static-pkg-config --libs glu` `i686-w64-mingw32.static-pkg-config --libs SDL2_net` -lm doom64ex-plus.res
