+++ src/engine/i_xinput.h
@@ -20,7 +20,7 @@
 //
 //-----------------------------------------------------------------------------
 
-#if defined(_WIN32) && defined(_USE_XINPUT)
+#if defined(_WIN32) && defined(USE_XINPUT)
 #ifndef __I_XINPUT__
 #define __I_XINPUT__
