+++ src/engine/m_keys.c
@@ -119,7 +119,7 @@ static keyinfo_t    Keys[] = {
 	{KEY_MWHEELDOWN,        "MouseWheelDown"},
 
 	// villsa 01052014
-#ifdef _USE_XINPUT  // XINPUT
+#if defined(_WIN32) && defined(USE_XINPUT)  // XINPUT
 	{BUTTON_DPAD_UP,        "DPadUp"},
 	{BUTTON_DPAD_DOWN,      "DPadDown"},
 	{BUTTON_DPAD_LEFT,      "DPadLeft"},
